+++ src/ca_tariff_parse/recognizers/__init__.py
@@ -12,6 +12,7 @@ from . import (
     proration,
     rate_table,
     sheet_rates,
+    transition_table,
 )
 
 __all__ = [
@@ -24,4 +25,5 @@ __all__ = [
     "proration",
     "rate_table",
     "sheet_rates",
+    "transition_table",
 ]